update draft and wire up upsert
Source: https://github.com/msupply-foundation/open-msupply/commit/ff486edbe6362778e48907ffa215be572a56e30f

+++ client/packages/common/src/types/schema.ts
@@ -1579,11 +1579,11 @@ export type DemographicProjectionNode = {
   __typename: 'DemographicProjectionNode';
   baseYear: Scalars['Int']['output'];
   id: Scalars['String']['output'];
-  year1: Scalars['Int']['output'];
-  year2: Scalars['Int']['output'];
-  year3: Scalars['Int']['output'];
-  year4: Scalars['Int']['output'];
-  year5: Scalars['Int']['output'];
+  year1: Scalars['Float']['output'];
+  year2: Scalars['Float']['output'];
+  year3: Scalars['Float']['output'];
+  year4: Scalars['Float']['output'];
+  year5: Scalars['Float']['output'];
 };
 
 export enum DemographicProjectionSortFieldInput {
@@ -2310,11 +2310,11 @@ export type InsertDemographicProjectionErrorInterface = {
 export type InsertDemographicProjectionInput = {
   baseYear: Scalars['Int']['input'];
   id: Scalars['String']['input'];
-  year1?: InputMaybe<Scalars['Int']['input']>;
-  year2?: InputMaybe<Scalars['Int']['input']>;
-  year3?: InputMaybe<Scalars['Int']['input']>;
-  year4?: InputMaybe<Scalars['Int']['input']>;
-  year5?: InputMaybe<Scalars['Int']['input']>;
+  year1?: InputMaybe<Scalars['Float']['input']>;
+  year2?: InputMaybe<Scalars['Float']['input']>;
+  year3?: InputMaybe<Scalars['Float']['input']>;
+  year4?: InputMaybe<Scalars['Float']['input']>;
+  year5?: InputMaybe<Scalars['Float']['input']>;
 };
 
 export type InsertDemographicProjectionResponse = DemographicProjectionNode | InsertDemographicProjectionError;
@@ -6908,11 +6908,11 @@ export type UpdateDemographicProjectionErrorInterface = {
 export type UpdateDemographicProjectionInput = {
   baseYear?: InputMaybe<Scalars['Int']['input']>;
   id: Scalars['String']['input'];
-  year1?: InputMaybe<Scalars['Int']['input']>;
-  year2?: InputMaybe<Scalars['Int']['input']>;
-  year3?: InputMaybe<Scalars['Int']['input']>;
-  year4?: InputMaybe<Scalars['Int']['input']>;
-  year5?: InputMaybe<Scalars['Int']['input']>;
+  year1?: InputMaybe<Scalars['Float']['input']>;
+  year2?: InputMaybe<Scalars['Float']['input']>;
+  year3?: InputMaybe<Scalars['Float']['input']>;
+  year4?: InputMaybe<Scalars['Float']['input']>;
+  year5?: InputMaybe<Scalars['Float']['input']>;
 };
 
 export type UpdateDemographicProjectionResponse = DemographicProjectionNode | UpdateDemographicProjectionError;
